fix(delete-user): DateEntity 수정으로 인한 match 및 기타 테스트 실패 해결
