+++ src/main/java/com/shootdoori/match/entity/User.java
@@ -3,8 +3,10 @@ package com.shootdoori.match.entity;
 import com.shootdoori.match.value.Password;
 import com.shootdoori.match.value.UniversityName;
 import jakarta.persistence.*;
+import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.security.crypto.password.PasswordEncoder;
 
+import java.time.LocalDateTime;
 import java.util.Objects;
 
 @Entity
@@ -50,6 +52,13 @@ public class User extends DateEntity {
     @Enumerated(EnumType.STRING)
     @Column(name = "POSITION", nullable = false, length = 2)
     private UserPosition position;
+
+    @Enumerated(EnumType.STRING)
+    @Column(nullable = false)
+    private UserStatus status = UserStatus.ACTIVE;
+
+    @LastModifiedDate
+    private LocalDateTime statusChangedAt;
 
     protected User() {
 
@@ -241,6 +250,18 @@ public class User extends DateEntity {
 
     public UserPosition getPosition() {
         return this.position;
+    }
+
+    public UserStatus getUserStatus() {
+        return status;
+    }
+
+    public void activate() {
+        this.status = UserStatus.ACTIVE;
+    }
+
+    public void requestDeletion() {
+        this.status = UserStatus.PENDING_DELETION;
     }
 
     public void update(String skillLevel, String position, String bio) {
